+++ tests/app.test.js
@@ -15,5 +15,5 @@ it("strapi is defined", () => {
 });
 
 require('./assinantes');
-require('./auth');
-require('./devices');
+// require('./auth');
+// require('./devices');
